模型表中的自增长ID 类型应为int(11)

+++ app/Application/Content/Data/Sql/cms_zhubiao.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `@cms@@zhubiao@` (
-  `id` mediumint(8) unsigned NOT NULL auto_increment,
+  `id` int(11) unsigned NOT NULL auto_increment,
   `catid` smallint(5) unsigned NOT NULL DEFAULT '0',
   `typeid` smallint(5) unsigned NOT NULL DEFAULT '0',
   `title` varchar(255) NOT NULL DEFAULT '',
